requirements.txt
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
idna==3.4
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
twilio==7.16.3
urllib3==1.26.14
wincertstore==0.2
